backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
pymongo==4.14.1
djongo==1.2.31
python-jose==3.5.0
cryptography==45.0.7
requests==2.32.5
openai==1.104.0
google-cloud-speech==2.33.0
Pillow==10.4.0
